// src/app/core/model/dpl-fail-header.ts
export class DplFailHeader {
  headerId: number;
  addressId: string;
  addressline: string;
  country: string;
  custNameC: string;
  custNameE: string;
  custNumber: string;
  docType: string;
  formNo: string;
  referenceNo: string;
  resultsStatus: string;
  resultsUrl: string;
  serviceType: string;
  tenant: string;
  trxDate: Date;
  lastUpdatedBy: string;
  createdBy: string;
  owner: string;
  ownerDept: string;
  lastUpdatedDate: Date;
  createdDate: Date;
  submitDate: Date;
  status: string;
  ouCode: string;
  ouName: string;
}